Adicao de uma pasta de views e o menu lateral
